Move project modules into actions app
This is a small repo, so no need for extra complexity
