+++ actions/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for registry project.
+ASGI config for actions project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -12,6 +12,6 @@ import os
 from django.core.asgi import get_asgi_application
 
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "registry.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "actions.settings")
 
 application = get_asgi_application()
