+++ actions/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for registry project.
+Django settings for actions project.
 
 Generated by 'django-admin startproject' using Django 3.2.5.
 
@@ -60,7 +60,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "registry.urls"
+ROOT_URLCONF = "actions.urls"
 
 TEMPLATES = [
     {
@@ -78,7 +78,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "registry.wsgi.application"
+WSGI_APPLICATION = "actions.wsgi.application"
 
 
 # Database
