+++ actions/templates/partials/header.html
@@ -4,7 +4,7 @@
     <div class="flex-1 flex">
       <div class="w-full flex lg:ml-0">
         <div class="relative w-full flex items-center">
-          <a class="text-base md:text-xl font-semibold text-white" href="{% url 'actions:index' %}">Actions</a>
+          <a class="text-base md:text-xl font-semibold text-white" href="{% url 'index' %}">Actions</a>
         </div>
       </div>
     </div>
