+++ actions/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for registry project.
+WSGI config for actions project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -12,6 +12,6 @@ import os
 from django.core.wsgi import get_wsgi_application
 
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "registry.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "actions.settings")
 
 application = get_wsgi_application()
