+++ pyproject.toml
@@ -30,4 +30,4 @@ skip_glob = [".direnv", "venv", ".venv"]
 use_parentheses = true
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "registry.settings"
+DJANGO_SETTINGS_MODULE = "actions.settings"
